djangoProject/core/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
pytz==2025.1
sqlparse==0.5.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
djangorestframework-simplejwt~=5.3.0
langchain_community==0.3.17
django_otp==1.5.4
qrcode==8.0
social-auth-app-django==5.4.3
psycopg2-binary>=2.9.9
django-sass-processor>=1.4
libsass==0.23.0
Pillow>=10.0.0
django-allauth==0.62.1
python-multipart==0.0.20